requirements.txt
streamlit
matplotlib
google-cloud-bigquery
google-auth
pandas
numpy
db-dtypes 
scikit-learn
joblib
google_generativeai
openai
google-cloud-secret-manager
google-auth-oauthlib
wordcloud
nltk
stopwords
PyPDF2
python-docx